[TM-3121] Refactor ScheduledJobsService to utilize MySQL named locks for task digest and polygon update email enqueuing; enhance test coverage and clean up email imports.
