+++ apps/job-service/src/scheduled-jobs/scheduled-jobs.service.ts
@@ -3,7 +3,7 @@ import { Cron, CronExpression } from "@nestjs/schedule";
 import { ScheduledJob, Task } from "@terramatch-microservices/database/entities";
 import { ENTERPRISES, LANDSCAPES } from "@terramatch-microservices/database/constants";
 import { AWAITING_APPROVAL, APPROVED } from "@terramatch-microservices/database/constants/status";
-import { Op, Transaction } from "sequelize";
+import { Op, QueryTypes, Transaction } from "sequelize";
 import {
   REPORT_REMINDER,
   SITE_AND_NURSERY_REMINDER,
@@ -15,13 +15,17 @@ import { Queue } from "bullmq";
 import { InjectQueue } from "@nestjs/bullmq";
 import { DateTime } from "luxon";
 import { REPORT_REMINDER_EVENT, SITE_AND_NURSERY_REMINDER_EVENT, TASK_DUE_EVENT } from "./scheduled-jobs.processor";
-import {
-  TaskDigestEmail,
-  WeeklyPolygonUpdateEmail
-} from "@terramatch-microservices/common/email/terrafund-report-reminder.email";
+import { TaskDigestEmail } from "@terramatch-microservices/common/email/terrafund-report-reminder.email";
+import { WeeklyPolygonUpdateEmail } from "@terramatch-microservices/common/email/weekly-polygon-update.email";
+import { batchFindAll } from "@terramatch-microservices/common/util/batch-find-all";
+import { PaginatedQueryBuilder } from "@terramatch-microservices/common/util/paginated-query.builder";
 
 const TASK_DIGEST_CHUNK_SIZE = 100;
 const POLYGON_DIGEST_CHUNK_SIZE = 50;
+
+/** MySQL/MariaDB named locks so only one job-service replica runs each weekly cron at a time. */
+const MYSQL_LOCK_TASK_DIGEST_WEEKLY = "tm_job_svc_task_digest_wk";
+const MYSQL_LOCK_POLYGON_UPDATES_WEEKLY = "tm_job_svc_polygon_update_wk";
 
 @Injectable()
 export class ScheduledJobsService {
@@ -116,48 +120,79 @@ export class ScheduledJobsService {
   }
 
   /**
-   * Laravel send-daily-digest-notifications: Monday 17:00 Europe/Sofia.
-   * Multiple job-service replicas may each fire this cron; use a single replica or a distributed
-   * lock if duplicate enqueues must be avoided (Laravel onOneServer).
+   * Weekly incomplete-task digest. Uses a MySQL named lock so only one replica enqueues when
+   * multiple job-service instances are deployed.
    */
-  @Cron("0 17 * * 1", { name: "taskDigestWeekly", timeZone: "Europe/Sofia" })
+  @Cron("0 17 * * 1", { name: "taskDigestWeekly" })
   async enqueueTaskDigestEmails(): Promise<void> {
-    this.logger.log("Enqueueing task digest email jobs (incomplete tasks)");
-    let cursor = 0;
-    let total = 0;
-    let hasMore = true;
-    while (hasMore) {
-      const tasks = await Task.findAll({
-        where: {
-          id: { [Op.gt]: cursor },
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    await Task.sequelize!.transaction(async transaction => {
+      await this.runWithMysqlNamedLock(MYSQL_LOCK_TASK_DIGEST_WEEKLY, transaction, async () => {
+        this.logger.log("Enqueueing task digest email jobs (incomplete tasks)");
+        let total = 0;
+        const builder = new PaginatedQueryBuilder(Task, TASK_DIGEST_CHUNK_SIZE).attributes(["id"]).where({
           status: { [Op.notIn]: [AWAITING_APPROVAL, APPROVED] }
-        },
-        attributes: ["id"],
-        order: [["id", "ASC"]],
-        limit: TASK_DIGEST_CHUNK_SIZE
+        });
+        for await (const page of batchFindAll(builder)) {
+          const taskIds = page.map(({ id }) => id);
+          total += taskIds.length;
+          await new TaskDigestEmail({ taskIds }).sendLater(this.emailQueue);
+        }
+        this.logger.log(`Task digest: enqueued chunks covering up to ${total} incomplete tasks`);
       });
-      if (tasks.length === 0) {
-        hasMore = false;
-        continue;
-      }
-      cursor = tasks[tasks.length - 1].id;
-      const taskIds = tasks.map(({ id }) => id);
-      total += taskIds.length;
-      await this.emailQueue.add(TaskDigestEmail.NAME, { taskIds });
-    }
-    this.logger.log(`Task digest: enqueued chunks covering up to ${total} incomplete tasks`);
+    });
   }
 
-  /** Laravel send-weekly-polygon-update-notifications: Monday 00:00 Europe/Sofia. */
-  @Cron("0 0 * * 1", { name: "weeklyPolygonUpdates", timeZone: "Europe/Sofia" })
+  /**
+   * Weekly polygon update digest. Uses a MySQL named lock so only one replica enqueues when
+   * multiple job-service instances are deployed.
+   */
+  @Cron("0 0 * * 1", { name: "weeklyPolygonUpdates" })
   async enqueueWeeklyPolygonUpdateEmails(): Promise<void> {
-    this.logger.log("Enqueueing weekly polygon update email jobs");
-    const weekAgo = DateTime.now().minus({ days: 7 }).toJSDate();
-    const uuids = await WeeklyPolygonUpdateEmail.loadRecentSitePolygonUuids(weekAgo);
-    for (let i = 0; i < uuids.length; i += POLYGON_DIGEST_CHUNK_SIZE) {
-      const sitePolygonUuids = uuids.slice(i, i + POLYGON_DIGEST_CHUNK_SIZE);
-      await this.emailQueue.add(WeeklyPolygonUpdateEmail.NAME, { sitePolygonUuids });
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    await Task.sequelize!.transaction(async transaction => {
+      await this.runWithMysqlNamedLock(MYSQL_LOCK_POLYGON_UPDATES_WEEKLY, transaction, async () => {
+        this.logger.log("Enqueueing weekly polygon update email jobs");
+        const weekAgo = DateTime.now().minus({ days: 7 }).toJSDate();
+        const uuids = await WeeklyPolygonUpdateEmail.loadRecentSitePolygonUuids(weekAgo);
+        for (let i = 0; i < uuids.length; i += POLYGON_DIGEST_CHUNK_SIZE) {
+          const sitePolygonUuids = uuids.slice(i, i + POLYGON_DIGEST_CHUNK_SIZE);
+          await new WeeklyPolygonUpdateEmail({ sitePolygonUuids }).sendLater(this.emailQueue);
+        }
+        this.logger.log(`Weekly polygon updates: enqueued ${uuids.length} UUIDs in chunks`);
+      });
+    });
+  }
+
+  /**
+   * Runs `fn` while holding a MySQL/MariaDB named lock on the current transaction connection.
+   * If the lock is not acquired (another node holds it), `fn` is skipped.
+   */
+  private async runWithMysqlNamedLock(
+    lockName: string,
+    transaction: Transaction,
+    fn: () => Promise<void>
+  ): Promise<void> {
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    const sequelize = Task.sequelize!;
+    const rows = await sequelize.query<{ got: number }>("SELECT GET_LOCK(:name, 0) AS got", {
+      replacements: { name: lockName },
+      transaction,
+      type: QueryTypes.SELECT
+    });
+    const first = rows[0];
+    if (first == null || Number(first.got) !== 1) {
+      this.logger.debug(`Cron skipped: could not acquire named lock [${lockName}]`);
+      return;
     }
-    this.logger.log(`Weekly polygon updates: enqueued ${uuids.length} UUIDs in chunks`);
+    try {
+      await fn();
+    } finally {
+      await sequelize.query("SELECT RELEASE_LOCK(:name) AS rel", {
+        replacements: { name: lockName },
+        transaction,
+        type: QueryTypes.SELECT
+      });
+    }
   }
 }
